Bug fix: disable the generator after close.
After the Accumulator.Close() method has been called, the data
structures used for the RNG do still exist and nothing prevents the
user from continuing using the generator after close.  This is a
problem, because the seed file is no longer periodically updated after
close.  To prevent this from causing problems, uninitialise the
underlying Generator when the Accumulator is closed.